+++ mongo-kubernetes-cluster3.yaml
@@ -1,0 +1,145 @@
+
+apiVersion: gateway.kyma.cx/v1alpha2
+kind: Api
+metadata:
+  name: personservice
+spec:
+# changeme: Replace {clusterhost} with cluster hostname (i.e. the part behind console in your kyma
+# console url
+  hostname: personservice.{clusterhost}
+  service:
+    name: personservice
+    port: 8080
+---
+apiVersion: v1
+kind: Service
+metadata:
+  labels:    
+     app: personservice
+     version: "0.0.3"
+  annotations:
+     auth.istio.io/80: NONE
+  name: personservice
+spec:
+  ports:
+  - name: http
+    port: 8080
+  selector:
+    app: personservice
+    version: "0.0.3"
+  type: NodePort
+
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+   annotations:
+      deployment.kubernetes.io/revision: "3"
+   labels:
+      run: personservice
+      mf_version: "2"
+   name: personservice   
+spec:
+   replicas: 2
+   selector:
+      matchLabels:
+         run: personservice
+   template:
+      metadata:
+         labels:    
+            app: personservice
+            run: personservice
+            version: "0.0.3"
+         annotations:
+            sidecar.istio.io/inject: "true"
+      spec:    
+         containers:
+           - name: personservice
+# changeme: Replace with pointer to your docker repository/container
+             image: "andy008/mongokubernetes:0.0.3"
+             imagePullPolicy: Always              
+             ports:
+              - containerPort: 8080
+                name: http
+              - containerPort: 8081
+                name: actuatorhttp
+                
+#             livenessProbe:
+#                httpGet:
+#                   path: /actuator/health
+#                   port: 8081
+#                initialDelaySeconds: 10
+#                timeoutSeconds: 40
+#                periodSeconds: 60
+#                failureThreshold: 3
+#             readinessProbe:
+#                httpGet:
+#                   path: /actuator/health
+#                   port: 8081
+#                periodSeconds: 60
+#                timeoutSeconds: 40
+#                initialDelaySeconds: 20
+#                failureThreshold: 3
+#                successThreshold: 2
+             env:
+              - name: spring_data_mongodb_host
+                valueFrom: 
+                   configMapKeyRef:
+                      name: mongo-kubernetes-config
+                      key: mongodb_host 
+              - name: spring_data_mongodb_database
+                valueFrom: 
+                   configMapKeyRef:
+                      name: mongo-kubernetes-config
+                      key: mongodb_database 
+              - name: spring_data_mongodb_authentication_database
+                valueFrom: 
+                   configMapKeyRef:
+                      name: mongo-kubernetes-config
+                      key: authentication_database 
+              - name: spring_data_mongodb_password
+                valueFrom: 
+                   secretKeyRef:
+                      name: first-mongo-mongodb
+                      key: mongodb-root-password 
+              - name: spring_data_mongodb_username
+                valueFrom: 
+                   configMapKeyRef:
+                      name: mongo-kubernetes-config
+                      key: mongodb_username
+              - name: personservicekubernetes_applicationconnetor_baseurl
+                valueFrom: 
+                   configMapKeyRef:
+                      name: mongo-kubernetes-config
+                      key: connector_baseurl
+              - name: spring_profiles_active
+                value: "ApplicationConnector,Cache"
+              - name: logging_level_com_sap_demo_service
+                value: "TRACE"
+              - name: spring_redis_host
+                value: "${#changeme}"   
+              - name: spring_redis_port
+                value: "${#changeme}"
+              - name: spring_redis_password
+                valueFrom: 
+                   secretKeyRef:
+                      name: #changeme
+                      key: redis-password            
+             volumeMounts:
+              - mountPath: /registration
+                name: registrationfile
+                readOnly: true
+              - mountPath: /security
+                name: kyma-certificate
+                readOnly: true
+         volumes:
+          - name: registrationfile
+            configMap:
+              name: registrationfile 
+              items:
+               - key: registrationfile.json
+                 path: registrationfile.json
+          - name: kyma-certificate
+            secret:
+              secretName: kyma-certificate
+
