+++ src/main/java/com/sap/demo/rest/RestErrorHandler.java
@@ -1,7 +1,5 @@
 package com.sap.demo.rest;
 
-import java.util.Map;
-
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.http.HttpStatus;
