+++ src/main/resources/application.properties
@@ -1,8 +1,13 @@
 spring.application.name=Person DB Kyma Sample App
 management.server.port=8081 
-logging.level.com.sap.demo=TRACE
-logging.level.org.apache.http.wire=TRACE
-#spring.profiles.active=ApplicationConnector
+#logging.level.com.sap.demo=TRACE
+#logging.level.org.apache.http.wire=TRACE
+spring.profiles.active=Cache
+
+# Redis Config (running on Docker)
+spring.cache.type=redis
+spring.redis.host=192.168.99.100
+spring.redis.port=6379
 
 
 personservicekubernetes.applicationconnetor.baseurl=https://gateway.playground.cluster.kyma.cx/personservicekubernetes
