+++ src/test/resources/application.properties
@@ -1,0 +1,3 @@
+spring.application.name=Person DB Kyma Sample App
+management.server.port=8081 
+spring.profiles.active=Default
